Move fixtures only used by integration tests into that directory
