+++ test/integration/fixtures/projects/containerized/inventory/hosts
