+++ test/integration/fixtures/projects/debug/inventory/inv_1
