+++ test/integration/fixtures/projects/debug/inventory/inv_2
