+++ test/integration/fixtures/projects/debug/project/debug.yml
