+++ test/integration/fixtures/projects/host_status/inventory
