+++ test/integration/fixtures/projects/host_status/project/gen_host_status.yml
