+++ test/integration/fixtures/projects/job_env/inventory/hosts
