+++ test/integration/fixtures/projects/printenv/inventory/hosts
