+++ test/integration/fixtures/projects/printenv/project/get_environment.yml
