+++ test/integration/fixtures/projects/sleep/inventory/hosts
