+++ test/integration/fixtures/projects/use_role/inventory/hosts
