+++ test/integration/fixtures/projects/use_role/roles/benthomasson.hello_role/meta/.galaxy_install_info
