+++ test/integration/fixtures/projects/use_role/roles/benthomasson.hello_role/meta/main.yml
