+++ test/integration/fixtures/projects/use_role/roles/benthomasson.hello_role/tasks/main.yml
