+++ test/integration/fixtures/projects/use_role/use_role.yml
